Put the test for HOST_DOWN or HOST_UP once again after the test for num_probes_active == 0 in HostScanStats::completed. The reason for this is fairly subtle and I didn't realize it at first: We have to make sure there are no active probes because once in the completed list, probes don't time out. Probes that are active stay active in the count. If the congestion window ever falls below the number of these active probes, the program will hang waiting for them to time out.
We could get away with this in the case of up hosts, because we call
HostScanStats::destroyAllOutstandingProbes in that case. We could do that in
the down case too, but that would prohibit a down host from being found up
later on. That's currently a matter of some luck; we don't keep sending probes
after a host is down but will accept replies to any other probes that have
already been sent.


git-svn-id: https://svn.nmap.org/nmap@11640 e0a8ed71-7df4-0310-8962-fdc924857419