Fix error: SCRIPT ENGINE (506.424s): ./scripts/pop3-capabilities.nse against a.b.1.47:995 ended with error: ./scripts/pop3-capabilities.nse:32: bad argument #1 to 'pairs' (table expected, got string)
which happens because pop3.lua returns a string error message instead of a table
of capabilities if it can't connect the socket or obtain a response from a
connected socket.
It now returns nil, err_message in these cases and the documentation now reflects
this - pop3-capabilities.nse silently returns and prints a debug message.

Added a 10s timeout for the socket in pop3.lua capabilities - 30s was a bit much.

git-svn-id: https://svn.nmap.org/nmap@11947 e0a8ed71-7df4-0310-8962-fdc924857419