Version detection updates:
Adjusted an Apache Stronghold matchline
Generated a softmatch for when Apache Stronghold does not present a version string.

Tweaked version detection for IdeaWebServer httpd to work against 302s and various 
response configurations.  Added matchline to grab X-Powered-By.

Added detection for IdeaPop3Server, IdeaSMTPServer, IdeaImapServer

Consolidated BIG-IP http / http-proxy related matchlines in a http-proxy matchline 
that takes into account various responses (302,404, etc).  Changed all instances 
of the name to be "BIG-IP" in order to be consistent with F5's documentation. 
The consolidated matchline is in the HTTP Options probe section as it appears to
be the most likely to generate a response from various versions of the platform.


git-svn-id: https://svn.nmap.org/nmap@23908 e0a8ed71-7df4-0310-8962-fdc924857419