Use getsockname to get source addresses.
This was previously gotten by setting the source address to be the same
of the interface address of the matching route. However this can be
wrong; when making a normal socket connection the source address is
chosen differently. We create a SOCK_DGRAM socket, connect it, and read
the local address with getsockname.

git-svn-id: https://svn.nmap.org/nmap@24413 e0a8ed71-7df4-0310-8962-fdc924857419