Honor sin6_scope_id in route_dst_generic.
This is set nonzero when there is a scope identifier at the end of
an IPv6 address, like fe80::a8bb:ccff:fedd:eeff%eth0. When this
happens, we look up the interface by index and then act as if it was the
interface given by -e. (But -e always has precedence over this.)

git-svn-id: https://svn.nmap.org/nmap@26242 e0a8ed71-7df4-0310-8962-fdc924857419