Use a sockaddr_storage for recvfrom in get_rpc_results.
Previously it was hardcoded to be sockaddr_in, which is obviously wrong
for IPv6. This was only used to filter out packets from other than the
host we are scanning. It may have still been succeeding by accident if
part of the IPv6 address had the bytes 00000000, because for me the port
number is at the same offset in sockaddr_in and sockaddr_in6, and
target->v4host().s_addr returns 00000000 for an IPv6 host.

git-svn-id: https://svn.nmap.org/nmap@26311 e0a8ed71-7df4-0310-8962-fdc924857419