Added support for LDAP substring searches to ldap.lua.  These can now be performed alone or in conjunction with other LDAP query types.
Added a new quick filter (qfilter) to ldap-search.nse that allows the user to specify, on the command line, an attribute and corresponding value to search the LDAP directory for.  The use of the asterisk '*' as a wildcard is permitted in the value parameter.

Updated asn1.lua with some minor notes on a hex value that was used.

git-svn-id: https://svn.nmap.org/nmap@26955 e0a8ed71-7df4-0310-8962-fdc924857419