Put the nselib/* loader at the front, not the end, of the loaders list.
The loaders list determines the order in which places are checked for
Lua modules. I had a problem on Windows when the nmap.exe directory
contained WinPcap's packet.dll. There was a runtime error when liblua
tried to open packet.dll instead of our packet.lua. This way we check
for our own files first.

git-svn-id: https://svn.nmap.org/nmap@28039 e0a8ed71-7df4-0310-8962-fdc924857419