Use ip_get_data_any, not ip_get_data, in ippackethdrinfo.
The difference is that ip_get_data_any doesn't require there to be an
upper-layer header at the end of an IPv6 chain. This avoids the message
"BOGUS!  Can't parse supposed IP packet" during -sO -6 scan. The -sO
sends non–upper-layer headers, some even with empty payloads, and so
causes this situation often.

git-svn-id: https://svn.nmap.org/nmap@28164 e0a8ed71-7df4-0310-8962-fdc924857419