Use getFlags16 for TCP flags in vectorize.
We use as features even the four reserved TCP flags (some of them are no
longer reserved but declared for a particular purpose by some RFCs).
getFlags only returns 8 bits, leading to garbage in the feature vector
in the higher-order positions.

This doesn't affect the integrity of fingerprint submissions, because
they copy the packet contents directly and don't go through the getFlags
accessor.

git-svn-id: https://svn.nmap.org/nmap@28236 e0a8ed71-7df4-0310-8962-fdc924857419