Remove byte order dependency in in_addr_to_octets.
Because in_addr is stored in network byte order, this function could
extract the octets in MSB-to-LSB order or in LSB-to-MSB order. This
didn't matter in the case of resolved names, because the same order was
used when generating the octet array and later when matching against it.
But the function parse_ipv4_ranges, which handles literal IPv4
addresses, always uses octet[0] as the MSB, so comparisons failed in
later matching.

As it was, the code worked on little-endian architectures but didn't on
big-endian.

git-svn-id: https://svn.nmap.org/nmap@28476 e0a8ed71-7df4-0310-8962-fdc924857419