Make it a warning (not a fatal error) when there's a problem with an IPv4 spec.
This is the way it worked for IPv6 (with the exception of a slash in an
IPv6 address, which has also been made into a warning)--warn about the
bogus specification and carry on with the other targets.

The main use for this is preventing the whole scan from being stopped
when a bad specification appears in a hostgroup after the first.

git-svn-id: https://svn.nmap.org/nmap@28515 e0a8ed71-7df4-0310-8962-fdc924857419