Update build instructions for MacPorts 2.1.1.
An important side effect of this change is the move the
--with-included-loaders flag from gtk2 to gdk-pixbuf2. This upstream
change:
http://git.gnome.org/browse/gtk+/commit/configure.in?h=gtk-2-24&id=cb29d2770714943af7b488a6a94f1f37b7466c8f
means that --with-included-loaders no longer has an effect when given to
gtk2's configure. This caused dynamic loaders to be searched for under
my /Users/david/macports-10.5 installation directory, which isn't
present after installing from the .dmg package.

Here's an report of a hang caused by the missing loaders:
http://seclists.org/nmap-dev/2012/q2/370

git-svn-id: https://svn.nmap.org/nmap@28771 e0a8ed71-7df4-0310-8962-fdc924857419